[RUSAA-1405] chore(control-api): fix CI failures — extract tests, sync openapi
- Extract test module from repos.rs into repos_tests.rs to bring repos.rs
  under the 600-line file-size cap (was 648, now 409 lines)
- Regenerate openapi.json to include the new 503 responses on
  POST /v1/repos/{id}/ingest (kafka_not_configured, kafka_unavailable)
- Regenerate frontend/src/api/generated/schema.ts from updated spec
- Fix PR body: replace tracker-only closes line with Closes #454 GitHub link
  and remove the secondary tracker reference that triggered the bundle check
